+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file, following t
 
 - Mol* core dependency updated to 4.5.0
   - Solves the bug with Export Models
+- Use BCIF files for AlphaFold models (unless `encoding: 'cif'`)
 - PDBeMolstarPlugin.extensions.foldseek
 - Added options `leftPanel`, `rightPanel`, `logPanel`, `tabs`
 - Option `hideCanvasControls` accepts "trajectory" value (for multi-model structures)